quitt-backend-v2/requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.24.1
flask-sqlalchemy==2.4.3
flask-marshmallow==0.13.0
marshmallow==3.6.1
numpy==1.18.5
regex==2020.6.8
webargs==6.1.0
Flask-Redis==0.4.0
jsonschema==3.2.0
pymysql==0.9.3
mysqlclient==1.4.6
celery==4.4.5
marshmallow-sqlalchemy==0.23.1
pytest==5.4.2
flask-sqlacodegen==1.1.7
pyinstaller==3.6
pypiwin32
joblib==0.16.0
scikit-learn==0.23.2
nltk==3.5